Move filtering logic to DynamicSchemaResolver
