+++ catalog/core/catalog-core-solr/src/main/java/ddf/catalog/source/solr/DynamicSchemaResolver.java
@@ -95,7 +95,7 @@ import org.slf4j.LoggerFactory;
  *
  * @since 0.2.0
  */
-public class DynamicSchemaResolver {
+public class DynamicSchemaResolver implements ConfigurationListener {
 
   private static final String LUX_XML_FIELD_NAME = "lux_xml";
 
@@ -163,6 +163,10 @@ public class DynamicSchemaResolver {
 
   Set<String> fieldsCache = new HashSet<>();
 
+  Set<String> anyTextFieldsCache = new HashSet<>();
+
+  Set<String> filteredAnyTextFieldsCache = new HashSet<>();
+
   private SchemaFields schemaFields;
 
   private Cache<String, MetacardType> metacardTypesCache =
@@ -173,8 +177,6 @@ public class DynamicSchemaResolver {
 
   private Processor processor = new Processor(new Config());
 
-  private ConfigurationStore configuration = ConfigurationStore.getInstance();
-
   public DynamicSchemaResolver(
       List<String> additionalFields, Function<TinyTree, TinyBinary> tinyBinaryFunction) {
     this(additionalFields);
@@ -182,6 +184,7 @@ public class DynamicSchemaResolver {
   }
 
   public DynamicSchemaResolver(List<String> additionalFields) {
+    ConfigurationStore.getInstance().addConfigurationListener(this);
     this.tinyBinaryFunction = this::newTinyBinary;
     this.schemaFields = new SchemaFields();
     metadataMaximumBytes = getMetadataSizeLimit();
@@ -191,7 +194,7 @@ public class DynamicSchemaResolver {
         Metacard.ID + SchemaFields.TEXT_SUFFIX + SchemaFields.TOKENIZED + SchemaFields.HAS_CASE);
     fieldsCache.add(Metacard.TAGS + SchemaFields.TEXT_SUFFIX);
 
-    configuration.addAnyTextField(Metacard.METADATA + SchemaFields.TEXT_SUFFIX);
+    anyTextFieldsCache.add(Metacard.METADATA + SchemaFields.TEXT_SUFFIX);
 
     fieldsCache.add(Validation.VALIDATION_ERRORS + SchemaFields.TEXT_SUFFIX);
     fieldsCache.add(Validation.VALIDATION_WARNINGS + SchemaFields.TEXT_SUFFIX);
@@ -200,6 +203,8 @@ public class DynamicSchemaResolver {
     fieldsCache.add(SchemaFields.METACARD_TYPE_OBJECT_FIELD_NAME);
 
     addAdditionalFields(this, additionalFields);
+
+    filterAnyTextFieldCache();
   }
 
   public DynamicSchemaResolver() {
@@ -227,7 +232,9 @@ public class DynamicSchemaResolver {
         .stream()
         .filter(descriptor -> BasicTypes.STRING_TYPE.equals(descriptor.getType()))
         .map(stringDescriptor -> stringDescriptor.getName() + SchemaFields.TEXT_SUFFIX)
-        .forEach(fieldName -> configuration.addAnyTextField(fieldName));
+        .forEach(fieldName -> anyTextFieldsCache.add(fieldName));
+
+    filterAnyTextFieldCache();
   }
 
   /**
@@ -275,9 +282,11 @@ public class DynamicSchemaResolver {
       String key = e.getKey();
       fieldsCache.add(key);
       if (key.endsWith(SchemaFields.TEXT_SUFFIX)) {
-        configuration.addAnyTextField(key);
+        anyTextFieldsCache.add(key);
       }
     }
+
+    filterAnyTextFieldCache();
   }
 
   /** Adds the fields of the Metacard into the {@link SolrInputDocument} */
@@ -747,7 +756,8 @@ public class DynamicSchemaResolver {
               + SchemaFields.HAS_CASE);
       fieldsCache.add(
           descriptor.getName() + schemaFields.getFieldSuffix(format) + SchemaFields.PHONETICS);
-      configuration.addAnyTextField(descriptor.getName() + schemaFields.getFieldSuffix(format));
+      anyTextFieldsCache.add(descriptor.getName() + schemaFields.getFieldSuffix(format));
+      filterAnyTextFieldCache();
     }
 
     if (format.equals(AttributeFormat.XML)) {
@@ -762,6 +772,57 @@ public class DynamicSchemaResolver {
               + schemaFields.getFieldSuffix(format)
               + getSpecialIndexSuffix(format));
     }
+  }
+
+  @Override
+  public void configurationUpdated() {
+    filterAnyTextFieldCache();
+  }
+
+  private void filterAnyTextFieldCache() {
+    Set<String> filteredList = new HashSet<>();
+
+    ConfigurationStore config = ConfigurationStore.getInstance();
+    List<String> anyTextFieldWhitelist = config.getAnyTextFieldWhitelist();
+    List<String> anyTextFieldBlacklist = config.getAnyTextFieldBlacklist();
+    if (!anyTextFieldBlacklist.isEmpty()) {
+      filteredList.addAll(anyTextFieldsCache);
+      for (String blacklistField : anyTextFieldBlacklist) {
+        String blacklist;
+        if (!blacklistField.endsWith(SchemaFields.TEXT_SUFFIX)) {
+          blacklist = blacklistField + SchemaFields.TEXT_SUFFIX;
+        } else {
+          blacklist = blacklistField;
+        }
+        filteredList.removeAll(
+            anyTextFieldsCache
+                .stream()
+                .filter(field -> field.matches(blacklist))
+                .collect(Collectors.toList()));
+      }
+    }
+
+    if (!anyTextFieldWhitelist.isEmpty()) {
+      for (String whitelistField : anyTextFieldWhitelist) {
+        String whitelist;
+        if (!whitelistField.endsWith(SchemaFields.TEXT_SUFFIX)) {
+          whitelist = whitelistField + SchemaFields.TEXT_SUFFIX;
+        } else {
+          whitelist = whitelistField;
+        }
+        filteredList.addAll(
+            anyTextFieldsCache
+                .stream()
+                .filter(field -> field.matches(whitelist))
+                .collect(Collectors.toList()));
+      }
+    }
+
+    if (anyTextFieldBlacklist.isEmpty() && anyTextFieldWhitelist.isEmpty()) {
+      filteredList.addAll(anyTextFieldsCache);
+    }
+
+    filteredAnyTextFieldsCache = filteredList;
   }
 
   private byte[] serialize(MetacardType anywhereMType) throws MetacardCreationException {
@@ -893,7 +954,7 @@ public class DynamicSchemaResolver {
   }
 
   Stream<String> anyTextFields() {
-    return configuration.getFilteredAnyTextFields().stream();
+    return filteredAnyTextFieldsCache.stream();
   }
 
   /**
